src/test/java/com/example/mvn1/controller/TestControllerTest.java
package com.example.mvn1.controller;

import com.example.mvn1.service.TestService;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.test.context.junit.jupiter.SpringExtension;
import org.springframework.test.web.servlet.MockMvc;

import static org.hamcrest.Matchers.is;
import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.when;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@ExtendWith(SpringExtension.class)
@WebMvcTest(TestController.class)
class TestControllerTest {

    @MockBean
    TestService testService;

    @Autowired
    private MockMvc mockMvc;


    @Test
    public void testIsPalindrome() throws Exception {
        when(testService.isPalindrome("xsx")).thenReturn(true);
        mockMvc.perform(get("/test/xsx"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.xsx",is(true)));
    }
    @Test
    public void testIsNotPalindrome() throws Exception {
        when(testService.isPalindrome("hello")).thenReturn(true);
        mockMvc.perform(get("/test/xsx"))
                .andDo(print())
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.xsx",is(false)));
    }

}